third_party/WebKit/LayoutTests/external/wpt/staticrange/idlharness-expected.txt

This is a testharness.js-based test.
PASS staticrange IDL
PASS StaticRange interface: existence and properties of interface object
FAIL StaticRange interface object length assert_equals: wrong value for StaticRange.length expected 1 but got 0
PASS StaticRange interface object name
PASS StaticRange interface: existence and properties of interface prototype object
PASS StaticRange interface: existence and properties of interface prototype object's "constructor" property
PASS StaticRange interface: existence and properties of interface prototype object's @@unscopables property
FAIL StaticRange interface: attribute startContainer assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
PASS Unscopable handled correctly for startContainer property on StaticRange
FAIL StaticRange interface: attribute startOffset assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
PASS Unscopable handled correctly for startOffset property on StaticRange
FAIL StaticRange interface: attribute endContainer assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
PASS Unscopable handled correctly for endContainer property on StaticRange
FAIL StaticRange interface: attribute endOffset assert_equals: setter must be function for PutForwards, Replaceable, or non-readonly attributes expected "function" but got "undefined"
PASS Unscopable handled correctly for endOffset property on StaticRange
PASS StaticRange interface: attribute collapsed
PASS Unscopable handled correctly for collapsed property on StaticRange
FAIL StaticRange interface: operation toRange() assert_own_property: interface prototype object missing non-static operation expected property "toRange" missing
PASS Unscopable handled correctly for toRange() on StaticRange
FAIL StaticRange must be primary interface of new StaticRange({ start: document.body, end: document.body }) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL Stringification of new StaticRange({ start: document.body, end: document.body }) assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL StaticRange interface: new StaticRange({ start: document.body, end: document.body }) must inherit property "startContainer" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL StaticRange interface: new StaticRange({ start: document.body, end: document.body }) must inherit property "startOffset" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL StaticRange interface: new StaticRange({ start: document.body, end: document.body }) must inherit property "endContainer" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL StaticRange interface: new StaticRange({ start: document.body, end: document.body }) must inherit property "endOffset" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL StaticRange interface: new StaticRange({ start: document.body, end: document.body }) must inherit property "collapsed" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
FAIL StaticRange interface: new StaticRange({ start: document.body, end: document.body }) must inherit property "toRange()" with the proper type assert_equals: Unexpected exception when evaluating object expected null but got object "TypeError: Illegal constructor"
Harness: the test ran to completion.
